Added function GetNumberOfRowsToShow() to determine the number of table rows needed. Only fights that are less than 3 weeks away are shown.
Added function WhereToInsertHeadings() to determine fight card heading locations. When a time difference of more than 12 hours is found between 2 fights then they belong to different events.

The http request for the API is still being ran when server.js is ran. It is in a helper function now but still only gets called when server.js is ran.

Removed a parameter in the LookupOdds() function and set the value of 'homeOrAway' team to a global variable that changes every time the function is ran. This works since home and away fighters will be alternating between who calls the function